--- client/src/lib/components/Icons/AnimatedBell/AnimatedBell.tsx
@@ -1,371 +1,0 @@
-import React, {forwardRef, useMemo} from 'react';
-import AnimatedLottieView from 'lottie-react-native';
-import AnimatedIcon, {AnimatedIconProps} from '../AnimatedIcon';
-import {COLORS} from '../../../../../../shared/src/constants/colors';
-
-const source = {
-  v: '5.8.1',
-  fr: 24,
-  ip: 0,
-  op: 72,
-  w: 300,
-  h: 300,
-  nm: 'Comp 2',
-  ddd: 0,
-  assets: [],
-  layers: [
-    {
-      ddd: 0,
-      ind: 1,
-      ty: 4,
-      nm: 'Layer 2',
-      parent: 2,
-      sr: 1,
-      ks: {
-        o: {a: 0, k: 100, ix: 11},
-        r: {a: 0, k: 0, ix: 10},
-        p: {
-          a: 1,
-          k: [
-            {
-              i: {x: 0.667, y: 1},
-              o: {x: 0.333, y: 0},
-              t: 0,
-              s: [0, 97.9, 0],
-              to: [-7.5, 0, 0],
-              ti: [-7.5, 0, 0],
-            },
-            {
-              i: {x: 0.667, y: 1},
-              o: {x: 0.333, y: 0},
-              t: 14,
-              s: [-45, 97.9, 0],
-              to: [7.5, 0, 0],
-              ti: [-3.167, 0, 0],
-            },
-            {
-              i: {x: 0.667, y: 1},
-              o: {x: 0.333, y: 0},
-              t: 25,
-              s: [45, 97.9, 0],
-              to: [3.167, 0, 0],
-              ti: [4.833, 0, 0],
-            },
-            {
-              i: {x: 0.667, y: 1},
-              o: {x: 0.333, y: 0},
-              t: 35,
-              s: [-26, 97.9, 0],
-              to: [-4.833, 0, 0],
-              ti: [-2, 0, 0],
-            },
-            {
-              i: {x: 0.667, y: 1},
-              o: {x: 0.333, y: 0},
-              t: 43,
-              s: [16, 97.9, 0],
-              to: [2, 0, 0],
-              ti: [0.333, 0, 0],
-            },
-            {
-              i: {x: 0.667, y: 1},
-              o: {x: 0.333, y: 0},
-              t: 48,
-              s: [-14, 97.9, 0],
-              to: [-0.333, 0, 0],
-              ti: [-0.667, 0, 0],
-            },
-            {
-              i: {x: 0.667, y: 1},
-              o: {x: 0.333, y: 0},
-              t: 54,
-              s: [14, 97.9, 0],
-              to: [0.667, 0, 0],
-              ti: [1.333, 0, 0],
-            },
-            {
-              i: {x: 0.667, y: 1},
-              o: {x: 0.333, y: 0},
-              t: 60,
-              s: [-10, 97.9, 0],
-              to: [-1.333, 0, 0],
-              ti: [-1.667, 0, 0],
-            },
-            {
-              i: {x: 0.667, y: 1},
-              o: {x: 0.333, y: 0},
-              t: 67,
-              s: [6, 97.9, 0],
-              to: [1.667, 0, 0],
-              ti: [1, 0, 0],
-            },
-            {t: 72, s: [0, 97.9, 0]},
-          ],
-          ix: 2,
-          l: 2,
-        },
-        a: {a: 0, k: [-0.5, 0, 0], ix: 1, l: 2},
-        s: {a: 0, k: [100, 100, 100], ix: 6, l: 2},
-      },
-      ao: 0,
-      shapes: [
-        {
-          ty: 'gr',
-          it: [
-            {
-              ind: 0,
-              ty: 'sh',
-              ix: 1,
-              ks: {
-                a: 0,
-                k: {
-                  i: [
-                    [-11.9, 0],
-                    [-0.4, 10],
-                    [0, 0],
-                  ],
-                  o: [
-                    [11.9, 0],
-                    [0, 0],
-                    [0.4, 9.9],
-                  ],
-                  v: [
-                    [-0.05, 9.5],
-                    [20.45, -9.5],
-                    [-20.45, -9.5],
-                  ],
-                  c: true,
-                },
-                ix: 2,
-              },
-              nm: 'Path 1',
-              mn: 'ADBE Vector Shape - Group',
-              hd: false,
-            },
-            {
-              ty: 'fl',
-              c: {a: 0, k: [1, 1, 1, 1], ix: 4},
-              o: {a: 0, k: 100, ix: 5},
-              r: 1,
-              bm: 0,
-              nm: 'Fill 1',
-              mn: 'ADBE Vector Graphic - Fill',
-              hd: false,
-            },
-            {
-              ty: 'tr',
-              p: {a: 0, k: [0, 0], ix: 2},
-              a: {a: 0, k: [0, 0], ix: 1},
-              s: {a: 0, k: [100, 100], ix: 3},
-              r: {a: 0, k: 0, ix: 6},
-              o: {a: 0, k: 100, ix: 7},
-              sk: {a: 0, k: 0, ix: 4},
-              sa: {a: 0, k: 0, ix: 5},
-              nm: 'Transform',
-            },
-          ],
-          nm: 'Group 1',
-          np: 2,
-          cix: 2,
-          bm: 0,
-          ix: 1,
-          mn: 'ADBE Vector Group',
-          hd: false,
-        },
-      ],
-      ip: 0,
-      op: 576,
-      st: 0,
-      bm: 0,
-    },
-    {
-      ddd: 0,
-      ind: 2,
-      ty: 4,
-      nm: 'Layer 1',
-      sr: 1,
-      ks: {
-        o: {a: 0, k: 100, ix: 11},
-        r: {
-          a: 1,
-          k: [
-            {
-              i: {x: [0.667], y: [1]},
-              o: {x: [0.333], y: [0]},
-              t: 0,
-              s: [0],
-            },
-            {
-              i: {x: [0.667], y: [1]},
-              o: {x: [0.333], y: [0]},
-              t: 7,
-              s: [-22],
-            },
-            {
-              i: {x: [0.667], y: [1]},
-              o: {x: [0.333], y: [0]},
-              t: 20,
-              s: [22],
-            },
-            {
-              i: {x: [0.667], y: [1]},
-              o: {x: [0.333], y: [0]},
-              t: 30,
-              s: [-10],
-            },
-            {
-              i: {x: [0.667], y: [1]},
-              o: {x: [0.333], y: [0]},
-              t: 39,
-              s: [6],
-            },
-            {
-              i: {x: [0.667], y: [1]},
-              o: {x: [0.333], y: [0]},
-              t: 46,
-              s: [-4],
-            },
-            {
-              i: {x: [0.833], y: [1]},
-              o: {x: [0.167], y: [0]},
-              t: 51,
-              s: [2],
-            },
-            {
-              i: {x: [0.833], y: [1]},
-              o: {x: [0.167], y: [0]},
-              t: 57,
-              s: [-2],
-            },
-            {t: 64, s: [0]},
-          ],
-          ix: 10,
-        },
-        p: {a: 0, k: [149.95, 57.7, 0], ix: 2, l: 2},
-        a: {a: 0, k: [-0.05, -77.2, 0], ix: 1, l: 2},
-        s: {a: 0, k: [100, 100, 100], ix: 6, l: 2},
-      },
-      ao: 0,
-      shapes: [
-        {
-          ty: 'gr',
-          it: [
-            {
-              ind: 0,
-              ty: 'sh',
-              ix: 1,
-              ks: {
-                a: 0,
-                k: {
-                  i: [
-                    [0, 0],
-                    [1.2, 12.4],
-                    [22.8, 8.2],
-                    [13.1, 0],
-                    [3.3, -11.7],
-                    [1.3, -33.1],
-                    [6.3, -6.1],
-                    [0, -9.5],
-                    [-9.2, 0],
-                    [0, 0],
-                    [0, 7.6],
-                    [8, 7.7],
-                  ],
-                  o: [
-                    [-6.2, -6.1],
-                    [-1.2, -33.1],
-                    [-3.4, -11.7],
-                    [-13.1, 0],
-                    [-22.9, 8.3],
-                    [-1.1, 12.4],
-                    [-8, 7.6],
-                    [0, 7.6],
-                    [0, 0],
-                    [9.1, 0],
-                    [0, -9.4],
-                    [0, 0],
-                  ],
-                  v: [
-                    [68.95, 39.5],
-                    [60.05, 8.9],
-                    [26.25, -56.6],
-                    [-0.05, -77.2],
-                    [-26.25, -56.6],
-                    [-60.15, 8.9],
-                    [-69.05, 39.5],
-                    [-85.55, 64.6],
-                    [-70.75, 77.2],
-                    [70.75, 77.2],
-                    [85.55, 64.6],
-                    [69.05, 39.5],
-                  ],
-                  c: true,
-                },
-                ix: 2,
-              },
-              nm: 'Path 1',
-              mn: 'ADBE Vector Shape - Group',
-              hd: false,
-            },
-            {
-              ty: 'fl',
-              c: {a: 0, k: [1, 1, 1, 1], ix: 4},
-              o: {a: 0, k: 100, ix: 5},
-              r: 1,
-              bm: 0,
-              nm: 'Fill 1',
-              mn: 'ADBE Vector Graphic - Fill',
-              hd: false,
-            },
-            {
-              ty: 'tr',
-              p: {a: 0, k: [0, 0], ix: 2},
-              a: {a: 0, k: [0, 0], ix: 1},
-              s: {a: 0, k: [100, 100], ix: 3},
-              r: {a: 0, k: 0, ix: 6},
-              o: {a: 0, k: 100, ix: 7},
-              sk: {a: 0, k: 0, ix: 4},
-              sa: {a: 0, k: 0, ix: 5},
-              nm: 'Transform',
-            },
-          ],
-          nm: 'Group 1',
-          np: 2,
-          cix: 2,
-          bm: 0,
-          ix: 1,
-          mn: 'ADBE Vector Group',
-          hd: false,
-        },
-      ],
-      ip: 0,
-      op: 576,
-      st: 0,
-      bm: 0,
-    },
-  ],
-  markers: [],
-};
-
-type AnimatedBellProps = Omit<AnimatedIconProps, 'source'>;
-
-export const AnimatedBell: React.FC<AnimatedBellProps> = forwardRef<
-  AnimatedLottieView,
-  AnimatedBellProps
->(({fill = COLORS.BLACK, ...props}, ref) => {
-  const colorFilters = useMemo(
-    () => [
-      {keypath: 'Layer 1', color: fill},
-      {keypath: 'Layer 2', color: fill},
-    ],
-    [fill],
-  );
-  return (
-    <AnimatedIcon
-      colorFilters={colorFilters}
-      ref={ref}
-      source={source}
-      {...props}
-    />
-  );
-});
