+++ client/src/routes/screens/Collection/Collection.tsx
@@ -36,7 +36,7 @@ import usePinnedCollectionById from '../../../lib/user/hooks/usePinnedCollection
 import {formatContentName} from '../../../lib/utils/string';
 import Markdown from '../../../lib/components/Typography/Markdown/Markdown';
 import AnimatedButton from '../../../lib/components/Buttons/AnimatedButton';
-import {AnimatedPlusToCheck} from '../../../lib/components/Icons';
+import {PlusToCheckIconAnimated} from '../../../lib/components/Icons';
 
 type Section = {
   title: string;
@@ -184,7 +184,7 @@ const Collection = () => {
                 small
                 onPress={togglePinned}
                 variant={isPinned ? 'primary' : 'secondary'}
-                AnimatedIcon={AnimatedPlusToCheck}
+                AnimatedIcon={PlusToCheckIconAnimated}
                 fill={COLORS.WHITE}
                 active={isPinned}>
                 {t('addToJourney')}
